requirements.txt
google-cloud-bigquery
google-oauth
google-auth
pandas
pandas-gbq
pyarrow
dotenv